Steam_TCS_database-realtime/requirements.txt
gradio==3.39.0
python-steam-api==1.0.8
pandas==2.0.0
requests==2.31.0
beautifulsoup4==4.12.2
matplotlib==3.7.1
openpyxl==3.1.2
